requirements.txt
tqdm==4.62.3
networkx==2.4
SQLAlchemy==1.4.25
six==1.13.0
pandas_profiling==3.0.0
scipy==1.6.0
numpy==1.20.1
matplotlib==3.4.3
pandas==1.2.3
psycopg2_binary==2.9.1
Jinja2==3.0.2
psycopg2==2.9.1
PyYAML==6.0
scikit_learn==1.0
